first round of refactoring without testing
